Refactoring on serial multi mux

+++ php/serial_multi_mux/serial_multi_mux.ino
@@ -10,13 +10,16 @@
 #define MAX 5
 #define VAL 6
 
-int units[TUNITS][7]= {
+
+byte units[TUNITS][7]= {
     {4, 5, 6, 7, 50, 58, 0},
     {10, 11, 12, 13, 60, 68, 0}
 };
 
-void setup() {
-    int i, j;
+
+void setup()
+{
+    byte i, j;
     for (i = 0; i < TUNITS; i++) {
         for (j = 0; j < 4; j++) {
             pinMode(units[i][j], OUTPUT);
@@ -25,7 +28,8 @@ void setup() {
     Serial.begin(9600);
 }
 
-void turn_on_led(int unit, int number) {
+void turn_on_led(byte unit, byte number)
+{
     //digitalWrite(units[unit][PWR], LOW);
     digitalWrite(units[unit][S0], number & 0x01);
     digitalWrite(units[unit][S1], number & 0x02); // no need to shift ;)
@@ -33,14 +37,16 @@ void turn_on_led(int unit, int number) {
     digitalWrite(units[unit][PWR], HIGH);
 }
 
-void turn_on_leds_on_unit(int unit) {
-    for (int i = 0; i < units[unit][VAL]; i++) {
+void turn_on_leds_on_unit(byte unit)
+{
+    for (byte i = 0; i < units[unit][VAL]; i++) {
         turn_on_led(unit, i);
     }
 }
 
-void set_actual(int value) {
-    for (int i = 0; i < TUNITS; i++) {
+void set_actual(byte value)
+{
+    for (byte i = 0; i < TUNITS; i++) {
         if (value >= units[i][MIN] && value <= units[i][MAX]) {
             units[i][VAL] = value - units[i][MIN];
             return;
@@ -48,12 +54,13 @@ void set_actual(int value) {
     }
 }
 
-void loop() {
+void loop()
+{
     if (Serial.available() > 0) {
-        int value = Serial.read();
+        byte value = Serial.read();
         set_actual(value);
     }
-    for (int i = 0; i < TUNITS; i++) {
+    for (byte i = 0; i < TUNITS; i++) {
         turn_on_leds_on_unit(i);
     }
 }
